agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore[strands-agents]>=1.4.7",
    "strands-agents>=1.30.0",
    "httpx>=0.28.0",
    "torch>=2.10",
    "scikit-learn>=1.8",
    "pandas>=3.0",
    "numpy>=2.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-httpx>=0.35.0",
    "respx>=0.22.0",
]

[project.scripts]
mars-agent = "src.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.6",
]

[tool.ruff.lint]
extend-select = ["I", "B", "UP"]
